src/requirements.txt
opencv-python
matplotlib
pandas
scipy-stack
scikit-image
pybalu
seaborn
python_speech_features
autopep8
spotipy
